Refactor NetworkAPI to be DSRestAPI
DSRestAPI short for DataStoreRestAPI is a better name
